solve SetFit import
